210406 meet category list 연동 구현 중 (#16) by DANA-HYEON
* Meet CRUD/Service 구현

모임 VO 구현
모임 CRUD구현
모임 Service 구현

* register ,get기능 구현

/meet/list.jsp html수정
register 기능 구현
get기능 구현

* meet update/delete

meet update/delete 기능 구현

* paging 구현

paging 구현 완료

* Search 구현 중

Search 구현 중

* meet/search기능 구현

meet/ search기능 구현

* meet / reply 기능 구현중

meet / reply 기능 구현중

* meet/reply 댓글 구현

reply 댓글 croud구현
대댓글 기능 구현 중

* meet/reply and category search

meet/reply 구현 중
meet/category search 구현 중

* meet/category 연동 구현 중

meet/category 연동 구현 중

+++ src/main/java/com/honeybee/domain/MeetReplyVO.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @Data
 public class MeetReplyVO {
 
-	private int mrno; //댓글번호
-	private int mno; //모임게시물번호
+	private Long mrno; //댓글번호
+	private Long mno; //모임게시물번호
 
 	private String id; //회원ID
 	private int grno2; //대댓글번호
